requirements.txt
# 🔥 Fire Detection IoT + ML Project Dependencies

# Serial communication with Arduino
pyserial==3.5

# Data manipulation and analysis
pandas==2.2.3
numpy==2.1.3

# Machine learning
scikit-learn==1.5.2
joblib==1.4.2

# Visualization (optional, for live graphs or dashboards)
matplotlib==3.9.2
seaborn==0.13.2

# Optional (for interactive dashboards / IoT extensions)
plotly==5.24.1
